// __generated__/ClubsListQuery.graphql.ts
export type order_by = "asc" | "asc_nulls_first" | "asc_nulls_last" | "desc" | "desc_nulls_first" | "desc_nulls_last" | "%future added value";
// ...
export type clubs_bool_exp = {
  _and?: ReadonlyArray<clubs_bool_exp> | null | undefined;
  _not?: clubs_bool_exp | null | undefined;
  _or?: ReadonlyArray<clubs_bool_exp> | null | undefined;
  contracts?: contracts_bool_exp | null | undefined;
  contracts_aggregate?: contracts_aggregate_bool_exp | null | undefined;
  country?: String_comparison_exp | null | undefined;
  created_at?: timestamp_comparison_exp | null | undefined;
  founded_year?: Int_comparison_exp | null | undefined;
  id?: uuid_comparison_exp | null | undefined;
  league?: String_comparison_exp | null | undefined;
  name?: String_comparison_exp | null | undefined;
  players?: players_bool_exp | null | undefined;
  players_aggregate?: players_aggregate_bool_exp | null | undefined;
  stadium?: String_comparison_exp | null | undefined;
  stats?: stats_bool_exp | null | undefined;
  stats_aggregate?: stats_aggregate_bool_exp | null | undefined;
  transfers?: transfers_bool_exp | null | undefined;
  transfersByToClubId?: transfers_bool_exp | null | undefined;
  transfersByToClubId_aggregate?: transfers_aggregate_bool_exp | null | undefined;
  transfers_aggregate?: transfers_aggregate_bool_exp | null | undefined;
  updated_at?: timestamp_comparison_exp | null | undefined;
};
// ...
export type clubs_order_by = {
  contracts_aggregate?: contracts_aggregate_order_by | null | undefined;
  country?: order_by | null | undefined;
  created_at?: order_by | null | undefined;
  founded_year?: order_by | null | undefined;
  id?: order_by | null | undefined;
  league?: order_by | null | undefined;
  name?: order_by | null | undefined;
  players_aggregate?: players_aggregate_order_by | null | undefined;
  stadium?: order_by | null | undefined;
  stats_aggregate?: stats_aggregate_order_by | null | undefined;
  transfersByToClubId_aggregate?: transfers_aggregate_order_by | null | undefined;
  transfers_aggregate?: transfers_aggregate_order_by | null | undefined;
  updated_at?: order_by | null | undefined;
};
// ...
export type ClubsListQuery$variables = {
  limit: number;
  offset: number;
  order_by?: ReadonlyArray<clubs_order_by> | null | undefined;
  where?: clubs_bool_exp | null | undefined;
};
export type ClubsListQuery$data = {
  readonly clubs: ReadonlyArray<{
    readonly country: string | null | undefined;
    readonly founded_year: number | null | undefined;
    readonly id: any;
    readonly league: string | null | undefined;
    readonly name: string;
    readonly stadium: string | null | undefined;
  }>;
  readonly clubs_aggregate: {
    readonly aggregate: {
      readonly count: number;
    } | null | undefined;
  };
};
export type ClubsListQuery = {
  response: ClubsListQuery$data;
  variables: ClubsListQuery$variables;
};